Improve Dockerfile readability
Use "here doc" syntax for RUN:

	RUN <<EOF
		..
		..
	EOF

Also use an array for installing the packages, so we don't need so many
\-escapes.

Both is this is to make things rather more readable.

As an aside, you don't need to use \-escapes at the end of line if it
already ends with &&, ||, or |. This just works:

	cmd1 &&
		cmd2

	cmd1 |
		cmd2

Also update the license metadata to AGPL3.